Changed the logic of remove_math_prefixes function
The function was removing math prefixes from any locate of string.

Example:

Bicampeão -> campeão
Taça Cambial -> Taça Camal

To solve this problem, I call regex library and define a pattern to
remove math prefixes only if in start of a string.
The char responsible for this is ^ in pattern regex method.